% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{add_log2_frac_to_numeric_vals}
\alias{add_log2_frac_to_numeric_vals}
\title{add_frac_to_numeric_vals}
\usage{
add_log2_frac_to_numeric_vals(tbl, frac)
}
\arguments{
\item{tbl}{}
}
\description{
The given numeric value will be added into all the numeric columns of the given tibble and resulted values will be log2
}
