% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{count_zero_vars}
\alias{count_zero_vars}
\title{Count variables having all values are 0 in a tbl.}
\usage{
count_zero_vars(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
an integer
}
\description{
Count variables having all values are 0 in a tbl.
}
\examples{
\dontrun{

 ggplot2::mpg \%>\% count_zero_vars()

 ###########################################

 tbl <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 )
 tbl \%>\% count_zero_vars()
}
}
