% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{get_zero_records}
\alias{get_zero_records}
\title{Subset records (rows) having all values are 0 in a tbl.}
\usage{
get_zero_records(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Subset records (rows) having all values are 0 in a tbl.
}
\examples{
\\donotrun{
 ggplot2::mpg \%>\% get_zero_records()

 ###########################################

 tbl <- tibble(x = c(0,1,2,0,3,5) , y = c(0,1,2,0,3,5) , z = c(0,1,2,0,3,5) )
 tbl
 tbl \%>\% get_zero_vars()

 tbl2 <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0  , xx = 0:4 , yy = 0)
 tbl2 \%>\% get_zero_vars()
}
}
