% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{remove_all_zero_rows}
\alias{remove_all_zero_rows}
\title{remove_all_zero_rows}
\usage{
remove_all_zero_rows(tbl)
}
\arguments{
\item{tbl}{}
}
\description{
remove_all_zero_rows
}
\details{
From the given tibble function removes all the rows in which all the numeric values are 0. Tibble of remaining rows will be returned.
}
