% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_convert_column_zscore}
\alias{tbl_convert_column_zscore}
\title{Scale numeric columns}
\usage{
tbl_convert_column_zscore(tbl, scale = TRUE, center = TRUE)
}
\arguments{
\item{tbl}{a tbl.}

\item{scale}{TRUE  look at \code{scale} argument of function base::scale().}

\item{center}{TRUE look at \code{center} argument of function base::scale().}
}
\value{
a tbl.
}
\description{
Scale numeric columns from a tbl. It uses \code{base::scale()} to scale the numeric vector.
}
\examples{
\dontrun{
tbl <- tibble(x = c(0,1,2,0,3,5) , y = c(0,1,2,0,3,5) , z = c(0,1,2,0,3,5) , zz = letters[1:6])
tbl \%>\% tbl_convert_column_zscore()
}


}
