% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{tbl_convert_log2}
\alias{tbl_convert_log2}
\title{Apply log2 on numeric variables}
\usage{
tbl_convert_log2(tbl, frac = 0)
}
\arguments{
\item{tbl}{a tbl}

\item{frac}{a numeric value to be added before applying log2}
}
\value{
a tbl
}
\description{
Apply log2 on numeric variables
}
\examples{
\dontrun{

 tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 )
 tbl
 tbl \%>\% tbl_convert_log2()
 tbl \%>\% tbl_convert_log2(frac = 0.01)

}
}
