% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_convert.R
\name{tbl_convert_row_zscore}
\alias{tbl_convert_row_zscore}
\title{Scale rows}
\usage{
tbl_convert_row_zscore(tbl, scale = TRUE, center = TRUE)
}
\arguments{
\item{tbl}{a tbl.}

\item{scale}{TRUE  for further details refer  \code{scale} argument of function \code{\link[base]{scale}}.
base::scale().}

\item{center}{TRUE for further details refer \code{center} argument of function \code{\link[base]{scale}}.}
}
\value{
a tbl.
}
\description{
Scale rows considering numeric columns from a tbl.
}
\examples{
\dontrun{
set.seed(12345)
tbl <- tibble(x = sample(c(1:10) , 5) , y = sample(c(1:10) , 5) , z = sample(c(1:10) , 5) , zz = letters[1:5])
tbl \%>\% tbl_convert_row_zscore()
}


}
\seealso{
\code{\link[base]{scale}}
}
