% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_count_vars_NA_any}
\alias{tbl_count_vars_NA_any}
\title{Count variables having atleast one NA}
\usage{
tbl_count_vars_NA_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
an integer
}
\description{
Count variables having atleast one NA
}
\examples{
\dontrun{


 tbl <- tibble(x = letters[1:5] , y = NA , z = c(1,2,3,NA,5) , xx= 1:5)
 tbl \%>\% tbl_count_vars_NA_any()
}
}
