% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_count_vars.R
\name{tbl_count_vars_zero_all}
\alias{tbl_count_vars_zero_all}
\title{Count columns / variables having all values are 0}
\usage{
tbl_count_vars_zero_all(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
an integer.
}
\description{
Count columns / variables having all values are 0 from a tbl.
}
\examples{
\dontrun{
 tbl <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 )
 tbl \%>\% tbl_count_vars_zero_all()
}
}
