% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_count_vars_zero_any}
\alias{tbl_count_vars_zero_any}
\title{Count columns having atleast one 0.}
\usage{
tbl_count_vars_zero_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
an integer
}
\description{
Count columns having atleast one 0.
}
\examples{
\dontrun{


 tbl <- tibble(x = letters[1:5] , y = NA , z = c(0,2,3,NA,5) , xx= 0:4)
 tbl \%>\% tbl_count_vars_zero_any()
}
}
