% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{tbl_count_zero_vars}
\alias{tbl_count_zero_vars}
\title{Count variables having all values are 0 in a tbl.}
\usage{
tbl_count_zero_vars(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
an integer
}
\description{
Count variables having all values are 0 in a tbl.
}
\examples{
\dontrun{


 tbl <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 )
 tbl \%>\% tbl_count_zero_vars()
}
}
