% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_extract_numbers}
\alias{tbl_extract_numbers}
\title{Extract numeric values from alpha-numeric values}
\usage{
tbl_extract_numbers(tbl, var)
}
\arguments{
\item{tbl}{a tbl.}

\item{var}{column name, the column fom which numbers are to be extracted}
}
\value{
a tbl
}
\description{
Extract numeric values from alpha-numeric values
}
\examples{
\dontrun{
alpha_num <-  cbind(paste0(LETTERS[1:10], 1:10), letters[1:10]) \%>\% as_tibble()
alpha_num  \%>\%  tbl_extract_numbers(V1)
}
}
