% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_extract_words}
\alias{tbl_extract_words}
\title{Extract words/alphabets from alpha-numeric values}
\usage{
tbl_extract_words(tbl, var)
}
\arguments{
\item{tbl}{a tbl}

\item{var}{column name, the column fom which words are to be extracted}
}
\value{
a tbl
}
\description{
Extract words/alphabets from alpha-numeric values
}
\examples{
\dontrun{
alpha_num <-  cbind(paste0(LETTERS[1:10], 1:10), letters[1:10]) \%>\% as_tibble()
alpha_num  \%>\%  tbl_extract_words(V1)
}

}
