% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_get_vars_NA_any}
\alias{tbl_get_vars_NA_any}
\title{Get columns names having atleast one NA.}
\usage{
tbl_get_vars_NA_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a cheracter vector contaning variable names
}
\description{
Get columns names having atleast one NA.
}
\examples{
\dontrun{


 tbl <- tibble(x = letters[1:5] , y = NA , z = c(1,2,3,NA,5) , xx= 1:5)
 tbl \%>\% tbl_get_vars_NA_any()
}
}
