% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_get_vars_zero_all}
\alias{tbl_get_vars_zero_all}
\title{Get variable (names) having all values are 0 in a tbl.}
\usage{
tbl_get_vars_zero_all(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a cheracter vector contaning variable names
}
\description{
Get variable (names) having all values are 0 in a tbl.
}
\examples{
\dontrun{

 tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 )
 tbl
 tbl \%>\% tbl_get_vars_zero_all()

 tbl2 <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0  , xx = 0:4 , yy = 0)
 tbl2 \%>\% tbl_get_vars_zero_all()
}
}
