% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_get_vars_zero_any}
\alias{tbl_get_vars_zero_any}
\title{Get variable (names) having atleast one 0}
\usage{
tbl_get_vars_zero_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a cheracter vector contaning variable names
}
\description{
Get variable (names) having atleast one 0
}
\examples{
\dontrun{

 tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 )
 tbl
 tbl \%>\% tbl_get_vars_zero_any()

 tbl2 <- tibble(x = letters[1:5] , y = NA , z = c(0,2,3,NA,5) , xx= 0:4)
 tbl2 \%>\% tbl_get_vars_zero_any()
}
}
