% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_keep_less_than_or_equal_all}
\alias{tbl_keep_less_than_or_equal_all}
\title{keep records with less than or equal \code{cutoff} in all numeric var}
\usage{
tbl_keep_less_than_or_equal_all(tbl, cutoff)
}
\arguments{
\item{tbl}{a tbl}

\item{cutoff}{numeric value to be used as cutoff}
}
\value{
a tbl
}
\description{
keep records with less than or equal \code{cutoff} in all numeric var
}
\examples{
\dontrun{
tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 , w = seq(1,10,by=2))
tbl \%>\% tbl_keep_less_than_or_equal_all(cutoff =4)
}

}
