% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_keep_rows_greater_than_or_equal_all}
\alias{tbl_keep_rows_greater_than_or_equal_all}
\title{Keep rows with greater than or equal 'cutoff' in all numeric columns.}
\usage{
tbl_keep_rows_greater_than_or_equal_all(tbl, cutoff)
}
\arguments{
\item{tbl}{a tbl}

\item{cutoff}{numeric value to be used as cutoff}
}
\value{
a tbl
}
\description{
Keep rows with greater than or equal 'cutoff' in all numeric columns.
}
\examples{
\dontrun{
tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 , w = seq(1,10,by=2))
tbl \%>\% tbl_keep_rows_greater_than_or_equal_all(cutoff =4)
}

}
