% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_keep_rows.R
\name{tbl_keep_rows_less_than_or_equal_any}
\alias{tbl_keep_rows_less_than_or_equal_any}
\title{Keep rows with less than or equal \code{'cutoff'} in any numeric column}
\usage{
tbl_keep_rows_less_than_or_equal_any(tbl, cutoff)
}
\arguments{
\item{tbl}{a tbl.}

\item{cutoff}{numeric value to be used as \code{'cutoff'}.}
}
\value{
a tbl.
}
\description{
Keep rows with less than or equal \code{'cutoff'} in any numeric column from a tbl.
}
\examples{
\dontrun{
tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 , w = seq(1,10,by=2))
tbl \%>\% tbl_keep_rows_less_than_or_equal_any(cutoff =4)
}

}
