% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_keep_rows.R
\name{tbl_keep_rows_NA_any}
\alias{tbl_keep_rows_NA_any}
\title{Subset rows having atleast one NA}
\usage{
tbl_keep_rows_NA_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Subset rows having atleast one NA across all columns / variables from a tbl.
}
\examples{
\dontrun{

 tbl <- tibble(x = c(0,1,2,0,3,NA) , y = c(0,1,NA,0,3,NA) , z = c(0,1,2,0,3,NA) )
 tbl
 tbl \%>\% tbl_keep_rows_NA_any()

 tbl2 <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0  , xx = 0:4 , yy = 0)
 tbl2 \%>\% tbl_keep_rows_NA_any()
}
}
