% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_keep_vars.R
\name{tbl_keep_vars_NA_all}
\alias{tbl_keep_vars_NA_all}
\title{Keep columns / variables having all values are NA}
\usage{
tbl_keep_vars_NA_all(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Keep columns / variables having all values are NA from a tbl. \code{tbl_keep_vars_NA_all()} takes care of both  numeric and non numeric columns.
}
\examples{
\dontrun{
 tbl <- tibble::tibble(x = letters[1:5] , y = NA , z = c(1,2,3,NA,5) )
 tbl \%>\% tbl_keep_vars_NA_all()
}

}
\seealso{
\code{\link[TidyWrappers]{tbl_remove_vars_NA_all}}
}
