% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_keep_vars.R
\name{tbl_keep_vars_NA_any}
\alias{tbl_keep_vars_NA_any}
\title{Keep columns / variables  having atlease one NA}
\usage{
tbl_keep_vars_NA_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Keep columns / variables  having atlease one NA from a tbl. \code{tbl_keep_vars_NA_any()} takes care of both  numeric and non numeric columns.
}
\examples{
\dontrun{
 tbl <- tibble::tibble(x = letters[1:5] , y = NA , z = c(1,2,3,NA,5) )
 tbl \%>\% tbl_keep_vars_NA_any()
}
}
\seealso{
\code{\link[TidyWrappers]{tbl_remove_vars_NA_any}}
}
