% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_keep_vars.R
\name{tbl_keep_vars_zero_any}
\alias{tbl_keep_vars_zero_any}
\title{Keep columns / variables having atleast one 0}
\usage{
tbl_keep_vars_zero_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Keep columns / variables having atleast one 0 from a tbl.
dev
}
\examples{
\dontrun{

 tbl <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 ,a = 0:4 , b = 1:5)
 tbl
 tbl \%>\% tbl_keep_vars_zero_any()

 tbl2 <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0  , xx = 0:4)
 tbl2 \%>\% tbl_keep_vars_zero_any()
}

}
\seealso{
\code{\link[TidyWrappers]{tbl_remove_vars_zero_any}}
}
