% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{tbl_remove_less_than_or_equal}
\alias{tbl_remove_less_than_or_equal}
\title{remove numeric values with less than \code{cutoff} in any one of the column}
\usage{
tbl_remove_less_than_or_equal(tbl, cutoff)
}
\arguments{
\item{tbl}{a tibble}

\item{cutoff}{numeric value to be used as cutoff}
}
\value{
a tibble
}
\description{
remove numeric values with less than \code{cutoff} in any one of the column
}
\examples{
\dontrun{
tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 , w = seq(1,10,by=2))
tbl \%>\% tbl_remove_less_than_or_equal(4)
}#'

}
