% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_remove_rows_zero_any}
\alias{tbl_remove_rows_zero_any}
\title{Remove rows having atleast one 0}
\usage{
tbl_remove_rows_zero_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Remove rows having atleast one 0 from a tbl.
}
\examples{

\dontrun{

 tbl <- tibble(x = c(0,1,2,0,3,5) , y = c(0,1,2,0,3,5) , z = c(0,1,2,0,3,5) )
 tbl
 tbl \%>\% tbl_remove_rows_zero_any()

 tbl2 <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 4:0  , xx = 0:4)
 tbl2 \%>\% tbl_remove_rows_zero_any()
}
}
