% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_remove_vars_zero_all}
\alias{tbl_remove_vars_zero_all}
\title{Remove columns having all values are 0.}
\usage{
tbl_remove_vars_zero_all(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Remove columns having all values are 0.
}
\examples{
\dontrun{


 tbl <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 )
 tbl
 tbl \%>\% tbl_remove_vars_zero_all()

 tbl2 <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0  , xx = 0:4)
 tbl2 \%>\% tbl_remove_vars_zero_all()
}
}
