% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_remove_vars_zero_any}
\alias{tbl_remove_vars_zero_any}
\title{Remove columns having atleast one 0}
\usage{
tbl_remove_vars_zero_any(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Remove columns having atleast one 0
}
\examples{
\dontrun{


 tbl <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 )
 tbl
 tbl \%>\% tbl_remove_vars_zero_any()

 tbl <- tibble(x = letters[1:5] , y = NA , z = c(0,2,3,NA,5) , xx= 0:4)
 tbl \%>\% tbl_remove_vars_zero_any()
}
}
