% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_remove_zero_vars}
\alias{tbl_remove_zero_vars}
\title{Remove variables having all values are 0 in a tbl.}
\usage{
tbl_remove_zero_vars(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a tbl.
}
\description{
Remove variables having all values are 0 in a tbl.
}
\examples{
\dontrun{


 tbl <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0 )
 tbl
 tbl \%>\% tbl_remove_zero_vars()

 tbl2 <- tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 0  , xx = 0:4)
 tbl2 \%>\% tbl_remove_zero_vars()
}
}
