% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_replace_greater_than}
\alias{tbl_replace_greater_than}
\title{replace numeric values greater than given \code{cutoff}}
\usage{
tbl_replace_greater_than(tbl, cutoff, replace_by)
}
\arguments{
\item{tbl}{a tbl.}

\item{cutoff}{a numeric value to be used as a cutoff.}

\item{replace_by}{a numeric value to be used to replace greater than \code{cutoff}}
}
\value{
a tbl.
}
\description{
replace numeric values greater than given \code{cutoff}
}
\examples{
\dontrun{
tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 , w = 5:1)
tbl \%>\% tbl_replace_greater_than(2, 1000 )
}

}
