% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_replace.R
\name{tbl_replace_less_than}
\alias{tbl_replace_less_than}
\title{Replace numeric values less than given \code{'cutoff'}}
\usage{
tbl_replace_less_than(tbl, cutoff, replace_by)
}
\arguments{
\item{tbl}{a tbl.}

\item{cutoff}{a numeric value to be used as a cutoff.}

\item{replace_by}{a numeric value to be used to replace less than \code{cutoff}.}
}
\value{
a tbl.
}
\description{
Replace numeric values less than given \code{'cutoff'} from a tbl.
}
\examples{
\dontrun{
tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 , w = 5:1)
tbl \%>\% tbl_replace_less_than(4, 0 )
}

}
