% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_replace.R
\name{tbl_replace_numeric}
\alias{tbl_replace_numeric}
\title{Replace numeric value in a tibble.}
\usage{
tbl_replace_numeric(tbl, value, replace_by)
}
\arguments{
\item{tbl}{a tbl.}

\item{value}{a numeric which is to be replaced from \code{tbl}.}

\item{replace_by}{a numeric replacement for \code{value}.}
}
\value{
a tbl
}
\description{
In a given tibble a numeric value can be replaced by another numeric value.
}
\examples{
\dontrun{
set.seed(1234)
tbl <- tibble::tibble(x = c(letters[1:5] ,letters[1:5]),
                       y = sample(1:5, 10 ,replace = T),
                       z = sample(1:5, 10,replace = T))
 tbl \%>\% tbl_replace_numeric(2, 1000)
 tbl \%>\% tbl_replace_numeric(4, 1000)
}

}
