% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{tbl_replace_string}
\alias{tbl_replace_string}
\title{Replace first occurance of \code{pattern} in all variables. Wrapper around \code{stringr::str_replace_all()}.}
\usage{
tbl_replace_string(tbl, pattern, replacement)
}
\arguments{
\item{tbl}{a tbl.}

\item{pattern}{pattern to look for. Will be passed to argument \code{pattern} to function \code{stringr::str_replace_all()}}

\item{replacement}{a character vector of replacements. Will be passed to argument \code{replacement} to function \code{stringr::str_replace_all()}}
}
\value{
a tbl
}
\description{
Replace first occurance of \code{pattern} in all variables. Wrapper around \code{stringr::str_replace_all()}.
}
\examples{

\dontrun{
 tbl <- tibble::tibble(x = c(letters[1:5] ,letters[1:5] ) ,
         y = c(LETTERS[1:5],LETTERS[1:5]) , z = 1:10 )
 tbl \%>\% tbl_replace_string(pattern = "a" , replacement = "a_changed")
 tbl \%>\% tbl_replace_string("[aeiou]", toupper)
 tbl \%>\% tbl_replace_string("([aeiou])", "")
 tbl \%>\% tbl_replace_string(c("a", "e", "i"), "-")
 tbl \%>\% tbl_replace_string("b", NA_character_)

}
}
\seealso{
stringr::str_replace_all()
}
