% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_convert_log}
\alias{tbl_convert_log}
\title{Apply log on numeric columns.}
\usage{
tbl_convert_log(tbl, frac = 0, base = 2)
}
\arguments{
\item{tbl}{a tbl}

\item{frac}{a numeric value to be added before applying log}

\item{base}{a positive number with respect to which log are computed}
}
\value{
a tbl
}
\description{
Apply log on numeric columns.
}
\examples{
\dontrun{

 tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 )
 tbl
 tbl \%>\% tbl_convert_log(base = 3)
 tbl \%>\% tbl_convert_log(base = 3 , frac = 0.01)

}
}
