% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_convert.R
\name{tbl_convert_log10}
\alias{tbl_convert_log10}
\title{Apply \code{log10()} on numeric columns / variables}
\usage{
tbl_convert_log10(tbl, frac = 0)
}
\arguments{
\item{tbl}{a tbl.}

\item{frac}{a numeric value to be added before applying log10.}
}
\value{
a tbl.
}
\description{
Apply \code{log10()} on numeric columns / variables from a tbl.
}
\examples{
\dontrun{

 tbl <- tibble::tibble(x = letters[1:5] , y = LETTERS[1:5] , z = 1:5 )
 tbl
 tbl \%>\% tbl_convert_log10()
 tbl \%>\% tbl_convert_log10(frac = 0.01)

}
}
