% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbl_get_vars_NA_all}
\alias{tbl_get_vars_NA_all}
\title{Get variables names having all values are NA in a tbl.}
\usage{
tbl_get_vars_NA_all(tbl)
}
\arguments{
\item{tbl}{a tbl.}
}
\value{
a cheracter vector contaning variable names
}
\description{
Get variables names having all values are NA in a tbl.
}
\examples{
\dontrun{


 tbl <- tibble(x = letters[1:5] , y = NA , z = c(1,2,3,NA,5) )
 tbl \%>\% tbl_get_vars_NA_all()
}
}
